feat(project-setup): add README.md, .gitignore, and LICENSE  Added the initial README.md to provide an overview and description of the project. Included a .gitignore file to manage files that should not be tracked by Git. Added the LICENSE file to define the terms under which the project can be used and distributed.
feat(project-setup): add README.md, .gitignore, and LICENSE

Added the initial README.md to provide an overview and description of the project.
Included a .gitignore file to manage files that should not be tracked by Git.
Added the LICENSE file to define the terms under which the project can be used and distributed.