feat(setup): Add mongoose connection and restructure controllers  Added mongoose to the exported connection in the connection.js so that the server will start. Changed the userControllers.js back to the original format of user-controllers.js and included route functions that are being exported. Added the index.js models export. Commented out routes in the userRoutes.js that aren't being used; however, they will be uncommented for use later. Established a database.  <footer>: This commit sets up the database connection and restructures the controllers and routes for better modularity.
feat(setup): Add mongoose connection and restructure controllers

Added mongoose to the exported connection in the connection.js so that the server will start. Changed the userControllers.js back to the original format of user-controllers.js and included route functions that are being exported. Added the index.js models export. Commented out routes in the userRoutes.js that aren't being used; however, they will be uncommented for use later. Established a database.

<footer>: This commit sets up the database connection and restructures the controllers and routes for better modularity.